Move difficulty adjustment code to separate module
