+++ frontend/src/app/dashboard/dashboard.component.scss
@@ -243,84 +243,6 @@
   max-width: 120px;
 }
 
-.difficulty-adjustment-container {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-around;
-  height: 76px;
-  .shared-block {
-    color: #ffffff66;
-    font-size: 12px;
-  }
-  .item {
-    padding: 0 5px;
-    width: 100%;
-    &:nth-child(1) {
-      display: none;
-      @media (min-width: 485px) {
-        display: table-cell;
-      }
-      @media (min-width: 768px) {
-        display: none;
-      }
-      @media (min-width: 992px) {
-        display: table-cell;
-      }
-    }
-  }
-  .card-text {
-    font-size: 22px;
-    margin-top: -9px;
-    position: relative;
-  }
-}
-
-
-.difficulty-skeleton {
-  display: flex;
-  justify-content: space-between;
-  @media (min-width: 376px) {
-    flex-direction: row;
-  }
-  .item {
-    max-width: 150px;
-    margin: 0;
-    width: -webkit-fill-available;
-    @media (min-width: 376px) {
-      margin: 0 auto 0px;
-    }
-    &:first-child{
-      display: none;
-      @media (min-width: 485px) {
-        display: block;
-      }
-      @media (min-width: 768px) {
-        display: none;
-      }
-      @media (min-width: 992px) {
-        display: block;
-      }
-    }
-    &:last-child {
-      margin-bottom: 0;
-    }
-  }
-  .card-text {
-    .skeleton-loader {
-      width: 100%;
-      display: block;
-      &:first-child {
-        margin: 14px auto 0;
-        max-width: 80px;
-      }
-      &:last-child {
-        margin: 10px auto 0;
-        max-width: 120px;
-      }
-    }
-  }
-}
-
 .loading-container {
   min-height: 76px;
 }
